refactor: All Connections Under One .csproj
Move all Exchanges to one proj
