+++ CryptoExchangeConstructs/Binance/BinanceMarketDataFeeds.cs
@@ -1,4 +1,4 @@
-namespace BinanceWebSocketConstructs;
+namespace CryptoExchangeConstructs.Binance;
 
 public static class BinanceMarketDataFeeds{
     public const string BaseEndpoint = "wss://data-stream.binance.vision/ws";
