+++ CryptoExchangeConstructs/Binance/BinanceMarketDataMethodTypes.cs
@@ -1,4 +1,4 @@
-namespace BinanceWebSocketConstructs;
+namespace CryptoExchangeConstructs.Binance;
 
 public static class BinanceMarketDataMethodTypes{
     public const string Subscribe = "SUBSCRIBE";
