+++ CryptoExchangeConstructs/Coinbase/CoinbaseMarketDataFeeds.cs
@@ -1,5 +1,4 @@
-namespace CoinbaseWebSocketConstructs;
-
+namespace CryptoExchangeConstructs.Coinbase;
 public static class CoinbaseMarketDataFeeds{
     public const string MarketDataEndpoint = "wss://advanced-trade-ws.coinbase.com";
     public const string UserOrderDataEndpoint = "wss://advanced-trade-ws-user.coinbase.com";
