+++ CryptoExchangeConstructs/Coinbase/CoinbaseMessageType.cs
@@ -1,6 +1,6 @@
-namespace KrakenWebSocketConstructs;
+namespace CryptoExchangeConstructs.Coinbase;
 
-public static class KrakenMethodTypes{
+public static class CoinbaseMessageTypes{
     public const string Subscribe = "subscribe";
     public const string Unsubscribe = "unsubscribe";
 }
