+++ CryptoExchangeConstructs/Coinbase/CoinbaseRequest.cs
@@ -1,4 +1,4 @@
-namespace CoinbaseWebSocketConstructs;
+namespace CryptoExchangeConstructs.Coinbase;
 
 using System.Security.Cryptography;
 using Newtonsoft.Json;
@@ -6,7 +6,7 @@ using AmazonSecretsManagerHandler;
 using SignatureHandling;
 using SignatureHandling.Interfaces;
 
-public class CoinbaseWebSocketSubscribeMessage
+public class CoinbaseRequest
 {
     [JsonProperty(PropertyName = "type")]
     public string Type { get; set; }
@@ -20,7 +20,7 @@ public class CoinbaseWebSocketSubscribeMessage
     [JsonProperty(PropertyName = "jwt")]
     public string Jwt { get; private set; }
     
-    public CoinbaseWebSocketSubscribeMessage(string type, string channel, string[] product_ids)
+    public CoinbaseRequest(string type, string channel, string[] product_ids)
     {
         Type = type;
         Channel = channel;
