+++ CryptoExchangeConstructs/Common/IExchangeRequest.cs
@@ -1,0 +1,5 @@
+namespace CryptoExchangeConstructs.Common;
+
+public interface IExchangeRequest{
+
+}
