+++ CryptoExchangeConstructs/Common/RequestFactory.cs
@@ -1,0 +1,1 @@
+namespace CryptoExchangeConstructs.Common;
