+++ CryptoExchangeConstructs/CryptoExchangeConstructs.csproj
@@ -6,13 +6,13 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
+    <ItemGroup>
    <PackageReference Include="Jose-JWT" Version="5.1.1" />
    <PackageReference Include="Microsoft.IdentityModel.Tokens" Version="8.8.0" />
    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
    <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="8.8.0" />
-   <ProjectReference Include="../../../AmazonSecretsManagerHandler/AmazonSecretsManagerHandler.csproj" />
-   <ProjectReference Include="../../../SignatureHandling/SignatureHandling.csproj" />
+   <ProjectReference Include="../AmazonSecretsManagerHandler/AmazonSecretsManagerHandler.csproj" />
+   <ProjectReference Include="../SignatureHandling/SignatureHandling.csproj" />
   </ItemGroup>
 
 </Project>
