+++ CryptoExchangeConstructs/Kraken/KrakenMethodTypes.cs
@@ -1,6 +1,7 @@
-namespace CoinbaseWebSocketConstructs;
+namespace CryptoExchangeConstructs.Kraken;
 
-public static class CoinbaseMessageTypes{
+
+public static class KrakenMethodTypes{
     public const string Subscribe = "subscribe";
     public const string Unsubscribe = "unsubscribe";
 }
