+++ CryptoExchangeConstructs/Kraken/KrakenRequest.cs
@@ -1,19 +1,20 @@
-namespace KrakenWebSocketConstructs;
+namespace CryptoExchangeConstructs.Kraken;
+
 
 using Newtonsoft.Json;
 
-public class KrakenMarketDataRequest{
+
+public class KrakenRequest
+{
     
     [JsonProperty(PropertyName = "method")]
     public string Method;
     [JsonProperty(PropertyName = "params")]
     public KrakenRequestParams Params;
 
-    public KrakenMarketDataRequest(string method, string[] symbol, string channel)
+    public KrakenRequest(string method, string[] symbol, string channel)
     {
         Method = method;
         Params = new(symbol, channel);
     }
-
-
 }
