+++ CryptoExchangeStreamer.sln
@@ -3,23 +3,17 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CoinbaseWebSocketConstructs", "CryptoExchangeConstructs\Coinbase\CoinbaseWebSocketConstructs\CoinbaseWebSocketConstructs.csproj", "{9F18DFF7-71D2-40CB-88F3-E4D20F3DE04D}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AmazonSecretsManagerHandler", "AmazonSecretsManagerHandler\AmazonSecretsManagerHandler.csproj", "{6A7428DF-F561-40E4-9147-6FB4D9DBFFF8}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SignatureHandling", "SignatureHandling\SignatureHandling.csproj", "{44ED1708-076E-4EEF-8134-37F5EEDE6360}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebsocketServer", "WebsocketServer\WebsocketServer.csproj", "{A4DC32C1-6FC4-4DEA-9491-7E1C07CB5C5F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BinanceWebSocketConstructs", "CryptoExchangeConstructs\Binance\BinanceWebSocketConstructs\BinanceWebSocketConstructs.csproj", "{69F4D3BC-ECBA-4C47-A0EC-83412E973523}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConnectionTester", "ConnectionTester\ConnectionTester.csproj", "{1C8B0D87-10DC-4480-B297-6359FC9963A2}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "CryptoExchangeConstructs", "CryptoExchangeConstructs", "{5C2565D0-B961-4657-6393-1FCCB545122A}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Kraken", "Kraken", "{E4C61BE3-27A7-32E9-E466-BCF011F9E5DA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KrakenWebSocketConstructs", "CryptoExchangeConstructs\Kraken\KrakenWebSocketConstructs\KrakenWebSocketConstructs.csproj", "{9830F41E-03F0-415A-9029-8C689E339B26}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CryptoExchangeConstructs", "CryptoExchangeConstructs\CryptoExchangeConstructs.csproj", "{C6F2037C-71C1-4EB8-A270-258C3B3AE104}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -115,6 +109,18 @@ Global
 		{9830F41E-03F0-415A-9029-8C689E339B26}.Release|x64.Build.0 = Release|Any CPU
 		{9830F41E-03F0-415A-9029-8C689E339B26}.Release|x86.ActiveCfg = Release|Any CPU
 		{9830F41E-03F0-415A-9029-8C689E339B26}.Release|x86.Build.0 = Release|Any CPU
+		{C6F2037C-71C1-4EB8-A270-258C3B3AE104}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C6F2037C-71C1-4EB8-A270-258C3B3AE104}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C6F2037C-71C1-4EB8-A270-258C3B3AE104}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C6F2037C-71C1-4EB8-A270-258C3B3AE104}.Debug|x64.Build.0 = Debug|Any CPU
+		{C6F2037C-71C1-4EB8-A270-258C3B3AE104}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C6F2037C-71C1-4EB8-A270-258C3B3AE104}.Debug|x86.Build.0 = Debug|Any CPU
+		{C6F2037C-71C1-4EB8-A270-258C3B3AE104}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C6F2037C-71C1-4EB8-A270-258C3B3AE104}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C6F2037C-71C1-4EB8-A270-258C3B3AE104}.Release|x64.ActiveCfg = Release|Any CPU
+		{C6F2037C-71C1-4EB8-A270-258C3B3AE104}.Release|x64.Build.0 = Release|Any CPU
+		{C6F2037C-71C1-4EB8-A270-258C3B3AE104}.Release|x86.ActiveCfg = Release|Any CPU
+		{C6F2037C-71C1-4EB8-A270-258C3B3AE104}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
