+++ WebsocketServer/ConnectionHandlers/ConnectionHandler
@@ -3,38 +3,47 @@ namespace WebsocketServer.ConnectionHandlers;
 using System.Text;
 using System.Net.WebSockets;
 using Newtonsoft.Json;
-using CryptoExchangeConstructs.Binance;
 
-public static class BinanceConnectionHandler{
-    public static async Task Connect(ClientWebSocket socket, string feed){
+public class ConnectionHandler{
+    
+    public ClientWebScoket Socket;
+    public string Source;
+
+    public ConnectionHandler(ClientWebSocket socket, string source)
+    {
+        Socket = socket;
+        Source = source;
+    }
+    
+    public async Task Connect(ClientWebSocket socket, string feed){
         Uri uri = new Uri(feed);
         await socket.ConnectAsync(uri, CancellationToken.None);
-        Console.WriteLine($"Connected to Binance WebSocket feed {feed}.");
+        Console.WriteLine($"Connected to {Source} WebSocket feed {feed}.");
     }
 
-    public static async Task Subscribe(ClientWebSocket socket, string channel, string[] product_ids)
+    public async Task Subscribe(ClientWebSocket socket, string channel, string[] product_ids)
     {
-        BinanceRequest request = new(channel, product_ids);
+        KrakenMarketDataRequest request = new(KrakenMethodTypes.Subscribe, product_ids, channel);
         string json = JsonConvert.SerializeObject(request);
-        Console.WriteLine("Sending Binance request message:");
+        Console.WriteLine($"Sending {Source} request message:");
         Console.WriteLine(json);
 
         var bytesToSend = new ArraySegment<byte>(Encoding.UTF8.GetBytes(json));
         await socket.SendAsync(bytesToSend, WebSocketMessageType.Text, true, CancellationToken.None);
     }
 
-    public static async Task Receive(ClientWebSocket socket, byte[] buffer)
+    public async Task Receive(ClientWebSocket socket, byte[] buffer)
     {
         var result = await socket.ReceiveAsync(new ArraySegment<byte>(buffer), CancellationToken.None);
         if (result.MessageType == WebSocketMessageType.Close)
         {
-            Console.WriteLine("Binance WebSocket closed.");
+            Console.WriteLine($"{Source} WebSocket closed.");
             await socket.CloseAsync(WebSocketCloseStatus.NormalClosure, string.Empty, CancellationToken.None);
         }
         else
         {
             string message = Encoding.UTF8.GetString(buffer, 0, result.Count);
-            Console.WriteLine("Received from Binance: " + message);
+            Console.WriteLine($"Received from {Source}: " + message);
 
             await RelayServer.BroadcastToClientsAsync(message);
         }
